requirements.txt
# ============================================
# Multimodal RCA Engine — Dependencies
# ============================================

# Core Data Science
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Machine Learning
scikit-learn>=1.3.0

# Log Parsing
drain3>=0.9.11
regex>=2023.0

# Dimensionality Reduction & Clustering
umap-learn>=0.5.3

# NLP & Text Processing
nltk>=3.8.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.0.0
ipywidgets>=8.0.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
requests>=2.31.0

# (Future) LLM Integration
# openai>=1.0.0
# google-generativeai>=0.3.0
# transformers>=4.35.0
# torch>=2.0.0
